+++ src/parser/index.ts
@@ -1,3 +1,13 @@
 export { parseSite } from './site-parser';
 export { extractMetadata } from './metadata-extractor';
-export type { ParsedSite, Heading, Link, Image, SiteMetadata, OpenGraphMetadata, TwitterCardMetadata } from './types';
+export { extractNavigation } from './navigation-extractor';
+export type {
+  ParsedSite,
+  Heading,
+  Link,
+  NavigationData,
+  Image,
+  SiteMetadata,
+  OpenGraphMetadata,
+  TwitterCardMetadata,
+} from './types';